move paging to JobDetailPage to simplify
